refactor(opx): add slice functions to their own package
