+++ opx/slicext/readme.go
@@ -1,0 +1,20 @@
+// Copyright (C) 2025 T-Force I/O
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// 	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+/*
+Package slicext implements helper functions specialized for slice
+that are not yet provided by Go lang standard libraries.
+Available since v0.3.0
+*/
+package slicext
